file reallocation to put the contents should be externally accessible under package
