+++ Cargo.toml
@@ -1,10 +1,47 @@
 [workspace]
-members = ["contracts/*"]
+members = [
+  "contracts/ibc-denom-resolver",
+  "contracts/packet-forward",
+  "packages/types",
+]
+resolver = "2"
+
+[workspace.package]
+version = "0.1.0"
+edition = "2021"
+
+# TODO: license
+
+# [profile.release]
+# rpath = false
+# lto = true
+# overflow-checks = true
+# opt-level = 3
+# debug = false
+# debug-assertions = false
+
 
 [profile.release]
-rpath = false
-lto = true
-overflow-checks = true
-opt-level = 3
-debug = false
+codegen-units    = 1
+debug            = false
 debug-assertions = false
+incremental      = false
+lto              = true
+overflow-checks  = true
+opt-level        = 3
+panic            = "abort"
+rpath            = false
+
+[workspace.dependencies]
+cosmwasm-std = "1.3.1"
+cosmwasm-schema = "1.3.1"
+thiserror = "1.0"
+cw-storage-plus = "1.1.0"
+cw-utils = "1.0.1"
+serde-json-wasm = "1.0.0"
+serde-cw-value = "0.7.0"
+prost = "0.11"
+prost-types = "0.11"
+
+# packages
+packet-forward-types = { path = "./packages/types" }
