+++ contracts/ibc-denom-resolver/Cargo.toml
@@ -1,15 +1,19 @@
 [package]
 name = "ibc-denom-resolver"
-version = "0.1.0"
-edition = "2021"
+version = {workspace = true}
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+doctest = false
+
 [dependencies]
-cosmwasm-std = { version = "1.3.1", features = ["stargate"] }
-cosmwasm-schema = "1.3.1"
-thiserror = "1.0"
-cw-storage-plus = "1.1.0"
-cw-utils = "1.0.1"
-serde-json-wasm = "0.5.1"
-packet-forward = { path = "../packet-forward" }
+cosmwasm-std = { workspace = true, features = ["stargate"] }
+cosmwasm-schema = {workspace = true}
+thiserror = {workspace = true}
+cw-storage-plus = {workspace = true}
+cw-utils = {workspace = true}
+serde-json-wasm = {workspace = true}
+packet-forward-types = {workspace = true}
