+++ contracts/ibc-denom-resolver/src/lib.rs
@@ -2,7 +2,4 @@ pub mod contract;
 pub mod error;
 pub mod execute;
 pub mod memo;
-pub mod msgs;
 pub mod query;
-pub mod state;
-pub mod types;
