+++ contracts/ibc-denom-resolver/src/query/config.rs
@@ -1,4 +1,4 @@
-use crate::{state::CONFIG, types::Config};
+use packet_forward_types::denom_resolver::{CONFIG, Config};
 use cosmwasm_std::{Deps, StdResult};
 
 #[cfg(not(feature = "library"))]
