+++ contracts/packet-forward/Cargo.toml
@@ -1,17 +1,22 @@
 [package]
 name = "packet-forward"
-version = "0.1.0"
-edition = "2021"
+version = {workspace = true}
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+doctest = false
+
 [dependencies]
-cosmwasm-std = { version = "1.3.1", features = ["stargate"] }
-cosmwasm-schema = "1.3.1"
-thiserror = "1.0"
-cw-storage-plus =  "1.1.0"
-cw-utils = "1.0.1"
-serde-json-wasm = "0.5.1"
-serde-cw-value = "0.7.0"
-prost = "0.11"
-prost-types = "0.11"
+cosmwasm-std = { wordspace = true, features = ["stargate"] }
+cosmwasm-schema = { wordspace = true}
+thiserror = { wordspace = true}
+cw-storage-plus =  { wordspace = true}
+cw-utils = { wordspace = true}
+serde-json-wasm = { wordspace = true}
+serde-cw-value = { wordspace = true}
+prost = { wordspace = true}
+prost-types = { wordspace = true}
+packet-forward-types = {workspace = true}
