+++ contracts/packet-forward/src/lib.rs
@@ -1,11 +1,7 @@
 pub mod contract;
 pub mod error;
 pub mod execute;
-pub mod ibc_hooks;
-pub mod msgs;
 pub mod proto;
 pub mod query;
 pub mod reply;
-pub mod state;
 pub mod sudo;
-pub mod types;
