+++ contracts/packet-forward/src/query/config.rs
@@ -1,4 +1,4 @@
-use crate::{state::CONFIG, types::Config};
+use packet_forward_types::packet_forward::{CONFIG, Config};
 use cosmwasm_std::{Deps, StdResult};
 
 #[cfg(not(feature = "library"))]
