+++ contracts/packet-forward/src/query/mod.rs
@@ -1,2 +1,3 @@
 pub mod config;
 pub mod failed_requests;
+pub mod ids;
