+++ packages/types/src/denom_resolver/mod.rs
@@ -1,0 +1,5 @@
+mod msgs;
+mod state;
+mod types;
+
+pub use self::{msgs::*, state::*, types::*};
