+++ packages/types/src/denom_resolver/state.rs
@@ -1,5 +1,5 @@
 use cw_storage_plus::Item;
 
-use crate::types::Config;
+use super::types::Config;
 
 pub const CONFIG: Item<Config> = Item::new("config");
